backend/validation-service/requirements.txt
fastapi==0.115.6
mangum==0.19.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
pydantic==2.10.3
python-dotenv==1.0.1
